Error-handling and divers config handling
Well I do not know if this is the best way to handle it but this commit prepares for working further. The next step will be to gather more unit tests and setting up the workflow for GitHub to further increase code stability.

TLDR: Sorry for the one reviewing this commit when I finish this branch